# Improve performance of segmentation