--- a/instant-segment-py/Cargo.toml
+++ b/instant-segment-py/Cargo.toml
@@ -16,7 +16,6 @@ name = "instant_segment"
 crate-type = ["cdylib"]
 
 [dependencies]
-ahash = "0.8"
 bincode = "1.3.2"
 instant-segment = { version = "0.10", path = "../instant-segment", features = ["with-serde"] }
 pyo3 = { version = "0.20", features = ["extension-module"] }