--- a/instant-segment-py/Cargo.toml
+++ b/instant-segment-py/Cargo.toml
@@ -2,6 +2,7 @@
 name = "instant-segment-py"
 version = "0.1.5"
 edition = "2018"
+rust-version = "1.65"
 license = "Apache-2.0"
 workspace = ".."
 description = "Fast English word segmentation"