This PR improves the speed of segmentation by pre-computing more values, using a different data structure for storing unigram and bigram data, and switching to a different hash function. In total, we achieve a 5-6x speed-up.

Here is what's happening:

1. [The first commit](https://github.com/instant-labs/instant-segment/commit/1aeee2129fff601b880f9130900cff7b9f52366d) just adds a second benchmark with a longer example. This is the performance we're starting with (on my machine):
```
test long  ... bench:     104,057 ns/iter (+/- 11,640)
test short ... bench:       4,902 ns/iter (+/- 439)
```

2. Then, in [Pre-compute unigram and bigram scores](https://github.com/instant-labs/instant-segment/commit/ea22d6cb654e159c52ac3681ac39170cd05b9312) we exploit the fact that `log(a * b) == log(a) + log(b)` and `log(a / b) == log(a) - log(b)` and apply this to the `score()` method. Once we've done that we see that we can pre-compute pretty much everything there by not storing the unigram and bigram counts in `Segmenter` but instead already the logarithms of the relative frequencies.
Result:
```
test long  ... bench:      86,257 ns/iter (+/- 8,231)
test short ... bench:       4,146 ns/iter (+/- 285)
```

3. [Use nested HashMap for storing both unigram and bigram scores](https://github.com/instant-labs/instant-segment/commit/dfee69080afcd8c22534b3c4fa080009e15feba9) is based on the observation that when we look up unigrams and bigrams in `score()` we have to calculate the hash value of `word` twice, and in cases where we do not have any unigram or bigram data for that particular word pair (the most common case) we actually unnecessarily compute hash values of 3 words (instead of just 1). By changing the data structure to a `HashMap<String, (f64, HashMap<String, f64>>` that maps a word to its unigram score & a nested HashMap in which the bigram score can be looked up using the previous word, we can reduce the amount of hashing needed quite a bit.
This will make construction of `Segmenter` more expensive, but since the intended usage pattern of this crate is to construct the `Segmenter` once and reuse it, this should be a worthwhile trade-off.
Result:
```
test long  ... bench:      19,651 ns/iter (+/- 2,097)
test short ... bench:       1,480 ns/iter (+/- 229)
```

5. Finally, we [Switch to rustc_hash instead of ahash](https://github.com/instant-labs/instant-segment/commit/6ed11a3562216339e835764c433f33b42ae9a2e1) as that performs better on small inputs, and, since we changed our data structure above, is now faster than `ahash`. I also tried `fnv` but it was slower than `rustc_hash` (but faster than `ahash`).
Final result:
```
test long  ... bench:      16,843 ns/iter (+/- 1,262)
test short ... bench:       1,081 ns/iter (+/- 223)
```

Note that 3. above is a breaking change, so will require a version-bump in the crate.